+++ iniesta/choices.py
@@ -2,7 +2,8 @@ from enum import IntFlag
 
 
 class InitializationTypes(IntFlag):
-    QUEUE_POLLING = 1  # 0001
-    EVENT_POLLING = 2  # 0010
 
-    SNS_PRODUCER = 16  # 10000
+    QUEUE_POLLING = 1  # 0001 = 1
+    EVENT_POLLING = 2  # 0010 = 2
+
+    SNS_PRODUCER = 16  # 10000 = 16